% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/boundary_layer_helpers.R
\name{add_asymmetry}
\alias{add_asymmetry}
\title{Add asymmetry due to storm motion}
\usage{
add_asymmetry(windspd, swd, mda, forward_speed)
}
\arguments{
\item{windspd}{Surface rotational wind speed at the location, in meters per
second.}

\item{swd}{Surface wind direction, in degrees. Due east is 0 degrees, due
north is 90 degrees, etc.}

\item{mda}{Bearing of the storm (i.e., direction of its forward motion),
in degrees. Due east is 0 degrees, due north is 90 degrees, etc.}

\item{forward_speed}{Forward motion of the storm, in meters per second}
}
\value{
Numeric vector with wind speed at grid location, with wind
   component due to storm motion added back in.
}
\description{
This function adds in a factor that accounts for asymmetry when adding back
in the forward motion of the storm to the wind speed at a location.
}
\note{
Only do this where windspd > 0 m/s. From NOAA Technical Report 23,
   Schwerdt et al., pg. 25
}
\examples{
add_asymmetry(windspd = 30, swd = 200, mda = 80, forward_speed = 10)
plot(add_asymmetry(windspd = 30, swd = 0:360, mda = 80, forward_speed = 10))

}
\references{
NOAA Technical Report NWS 23, Schwerdt and Watkins, 1979.
}

