% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{add_forward_speed}
\alias{add_forward_speed}
\title{Add in forward speed of the storm}
\usage{
add_forward_speed(wind_sfc_sym, tcspd_u, tcspd_v, swd, cdist, Rmax)
}
\arguments{
\item{wind_sfc_sym}{A numeric value with the estimated symmetric surface
wind speed at the grid point, in meters / second}

\item{tcspd_u}{A numeric value with the u-component of the tropical cyclone
speed in meters / second}

\item{tcspd_v}{A numeric value with the v-component of the tropical cyclone
speed in meters / second}

\item{swd}{A numeric value with the surface wind direction in degrees}

\item{cdist}{A numeric value with the radius from the storm's center to the
grid point in kilometers}

\item{Rmax}{A numeric value with the radius at which maximum winds occur in kilometers}
}
\value{
windspeed A numeric value with the asymmetric surface windspeed at the
location, in meters/second
}
\description{
Adds the storm's forward speed component (i.e., motion asymmetry) back
into the estimated surface wind speed at a grid point location after
rotational winds have been modeled for the location.
}
\details{
This function uses equation 12 from Phadke et al. (2003).
}
\references{
Phadke AC, Martino CD, Cheung KF, and Houston SH. 2003. Modeling of
   tropical cyclone winds and waves for emergency management. Ocean
   Engineering 30(4):553-578.
}
