% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{add_inflow_new}
\alias{add_inflow_new}
\title{Add inflow to direction of surface winds}
\usage{
add_inflow_new(gwd, cdist, Rmax, tclat, glandsea)
}
\arguments{
\item{gwd}{A numeric value with the gradient wind direction in degrees}

\item{cdist}{A numeric value with the radius from the storm's center to the
grid point in kilometers}

\item{Rmax}{A numeric value with the radius at which maximum winds occur in kilometers}

\item{tclat}{A numeric value with latitude in radians}
}
\value{
swd A numeric value with the surface wind direction in degrees
}
\description{
Add inflow to direction of surface winds
}
