% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/my_wind_code.R
\name{add_storm_motion_wind}
\alias{add_storm_motion_wind}
\title{Add back in wind component due to storm motion}
\usage{
add_storm_motion_wind(windspd, swd, mda, forward_speed)
}
\value{
Numeric vector with wind speed at grid location, with wind
   component due to storm motion added back in
}
\description{
Add back in wind component due to storm motion
}
\note{
Only do this where windspd > 0 m/s. From NOAA Technical Report 23,
   Schwerdt et al., pg. 25
}

