% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/map_winds.R
\name{add_storm_track}
\alias{add_storm_track}
\title{Plot Atlantic basin hurricane tracks}
\usage{
add_storm_track(storm_tracks, plot_object, padding = 2, plot_points = FALSE,
  alpha = 1, color = "firebrick")
}
\arguments{
\item{storm_tracks}{Character vector with the names of all storms to plot.
This parameter must use the unique storm identifiers from the
`storm_id` column of the `hurr_tracks` dataframe.}

\item{plot_object}{NULL or the name of a ggplot object to use as the
underlying plot object. If NULL, the function will generate a new
map of the eastern US states using `default_map`.}

\item{padding}{Numerical value giving the number of degrees to add to the
outer limits of the plot object (or default map if `plot_object` is
left as NULL) when cropping hurricane tracks.}

\item{plot_points}{TRUE / FALSE indicator of whether to include points,
as well as lines, when plotting the hurricane tracks.}

\item{alpha}{Numerical value designating the amount of transparency to
use for plotting tracks.}

\item{color}{Character string giving the color to use to plot the tracks.}
}
\value{
Returns a ggplot object with plotting data for the storm tracks
   of the selected storms. This object can be printed directly or added
   on to with other ggplot commands.
}
\description{
Plot the tracks of any selected storms in the hurricane tracking
   dataset for the Atlantic basin. This function allows you to
   plot a new map or add the tracks to an existing ggplot object.
}
\examples{
\dontrun{
library(ggplot2)
data("county_points")
grid_winds_floyd <- get_grid_winds(hurr_track = floyd_tracks,
                                   grid_df = county_points)
floyd_map <- map_wind(grid_winds_floyd, value = "vmax_sust",
                      wind_metric = "knots") +
             ggtitle("Maximum sustained wind speeds")
add_storm_track(floyd_tracks, plot_object = floyd_map)
}

}

