% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/my_wind_code.R
\name{calc_and_summarize_grid_wind}
\alias{calc_and_summarize_grid_wind}
\title{Calculate and summarize grid winds}
\usage{
calc_and_summarize_grid_wind(grid_point = stormwindmodel::county_points[1, ],
  with_wind_radii = add_wind_radii(), tint = 0.25, gust_duration_cut = 20,
  sust_duration_cut = 20)
}
\description{
This function combines the `calc_grid_wind` and `summarize_grid_wind`
functions so they can be run jointly in the overall `get_grid_winds`
function.
}

