% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/latlon_helpers.R
\name{calc_bearing}
\alias{calc_bearing}
\title{Calculate the direction from one location to another}
\usage{
calc_bearing(phi_1, L_1, phi_2, L_2)
}
\arguments{
\item{phi_1}{Latitude of the first location, in degrees}

\item{L_1}{Longitude of the first location, in degrees}

\item{phi_2}{Latitude of the second location, in degrees}

\item{L_2}{Longitude of the second location, in degrees}
}
\value{
The direction of the second location as seen from the first location,
   in [units]. For example, 0 degrees if the second location is due west of the
   first, and 90 degrees if the second location is due north of the first.
}
\description{
This function calculates the bearing of a second location, as seen from
the first location, based on latitude and longitude coordinates for both
locations.
}

