% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{calc_bearing_single}
\alias{calc_bearing_single}
\title{Calculate bearing from one lat/long to another for a single point}
\usage{
calc_bearing_single(tclat, tclon, glat, glon)
}
\arguments{
\item{tclat}{A numeric value with the latitude of the tropical cyclone's center
in radians}

\item{tclon}{A numeric value with the longitude of the tropical cyclone's center
in radians}

\item{glat}{A numeric value with the latitude of the grid point in radians}

\item{glon}{A numeric value with the longitude of the grid point in radians}
}
\value{
A numeric value with the bearing from the storm's center to the grid point
  in polar coordinates
}
\description{
Calculate bearing from one lat/long to another for a single point
}
