% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/boundary_layer_helpers.R
\name{calc_corr_fct}
\alias{calc_corr_fct}
\title{Calculate the correction factor to add forward speed to modeled surface wind}
\usage{
calc_corr_fct(Rmax, cdist)
}
\arguments{
\item{Rmax}{A numeric vector giving the radius to maximum winds (in kilometers)
for the tropical storm.}

\item{cdist}{A numeric vector giving radius (in kilometers) from the storm
center to the location being modeled.}
}
\value{
A numeric vector giving the correction factor, calculated based
  on Phadke et al. 2003
}
\description{
Calculates the correction factor (Phadke et al. 2003) to use to adds the
storm's forward speed component (i.e., motion asymmetery) back
into the estimated u- and v-components of the surface wind speed at a grid point
location after u- and v-components of rotational winds have been modeled for the
location.
}
\details{
This function uses equation 12 from Phadke et al. (2003).
}
\references{
Phadke AC, Martino CD, Cheung KF, and Houston SH. 2003. Modeling of
   tropical cyclone winds and waves for emergency management. Ocean
   Engineering 30(4):553-578.
}
