% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{calc_distance}
\alias{calc_distance}
\title{Calculate distance between two latitude / longitude points}
\usage{
calc_distance(tclat, tclon, glat, glon, Rearth = 6371)
}
\arguments{
\item{tclat}{A numeric value with the latitude of the tropical cyclone's center
in radians}

\item{tclon}{A numeric value with the longitude of the tropical cyclone's center
in radians}

\item{glat}{A numeric value with the latitude of the grid point in radians}

\item{glon}{A numeric value with the longitude of the grid point in radians}

\item{Rearth}{A numeric value with the radius of the earth in kilometers. This
value defaults to 6,371 km, which is the median radius of the earth.}
}
\value{
A numeric value with the distance (in kilometers) from the tropical
  cyclone center to the grid point.
}
\description{
Calculate distance between two latitude / longitude points
}
