% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{calc_forward_speed}
\alias{calc_forward_speed}
\title{Calculate storm's forward speed}
\usage{
calc_forward_speed(tclat_1, tclon_1, time_1, tclat_2, tclon_2, time_2)
}
\arguments{
\item{time_1}{A date-time vector giving the time of the first observation.}

\item{time_2}{A date-time vector giving the time of the second observation.}
}
\value{
A numeric vector with the average forward speed of the storm between
   the two observations, in meters per second.
}
\description{
This storm takes two storm locations and their observations times and
calculates the average speed of the storm between the two observations.
}
