% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{calc_gwd}
\alias{calc_gwd}
\title{Calculate gradient wind direction at a point}
\usage{
calc_gwd(tclat, tclon, glat, glon)
}
\arguments{
\item{tclat}{A numeric value with the latitude of the tropical cyclone's center
in radians}

\item{tclon}{A numeric value with the longitude of the tropical cyclone's center
in radians}

\item{glat}{A numeric value with the latitude of the grid point in radians}

\item{glon}{A numeric value with the longitude of the grid point in radians}
}
\value{
A numeric value with the the direction of gradient storm winds at a location,
  in polar coordinates (i.e., due East is 0 degrees, due North is 90 degrees, etc.)
}
\description{
This function determines the bearing from the storm's center to a
grid point location and then calculates the direction of gradient
winds at that grid point. Gradient winds will be perpendicular to
the direction from the storm's center to the grid point. In the
Northern hemisphere, cyclonic winds are counterclockwise, so you
add 90 degrees to the bearing from the storm center to the grid point
when using polar coordinates. In the Southern hemisphere, cyclonic
winds are clockwise, so you subtract 90 degrees.
}
