% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{calc_linear_coefs}
\alias{calc_linear_coefs}
\title{Calculate coefficients for linear interpolation}
\usage{
calc_linear_coefs(x, y)
}
\arguments{
\item{x}{A numeric vector with independent values of the original function.}

\item{y}{A numeric vector with dependent values of the original function.}
}
\value{
A numeric vector with slope coefficients for the linear interpolation of
  the original function. This vector has a length one less that the input vectors.
  The slope is for the section starting at that location in the x input vector.
}
\description{
Calculate coefficients for linear interpolation
}
