% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/willoughby2006eqs.R
\name{calc_R1}
\alias{calc_R1}
\title{Calculate R1 using Rmax and xi}
\usage{
calc_R1(Rmax, xi)
}
\arguments{
\item{Rmax}{Numeric vector of the radius at which the maximum wind occurs,
in kilometers}

\item{xi}{A numeric value with the \eqn{\xi} value determined by
\code{\link{solve_for_xi}}}
}
\value{
A numeric vector with the estimated value of \eqn{R_1} for the
   storm track observation.
}
\description{
Once you've solved for \eqn{xi}, use this value and the estimated
\eqn{R_{max}} to determine the radius from the storm center to the start of
the transition region.
}

