% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/boundary_layer_helpers.R
\name{calc_sfc_final}
\alias{calc_sfc_final}
\title{Calculate final surface wind direction}
\usage{
calc_sfc_final(wind_sfc_u, wind_sfc_v)
}
\arguments{
\item{wind_sfc_u}{A numeric vector with the tropical cyclone speed, u-component
(m / s).}
}
\value{
A numeric vector giving the final surface wind direction,
 including both symmetrical and forward storm motion components of
 the wind.
}
\description{
Determines the final direction of the wind at a given grid point
for a given storm location, based on both the symmetrical storm
winds and the forward motion of the storm.
}
