% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/boundary_layer_helpers.R
\name{calc_sfc_u}
\alias{calc_sfc_u}
\title{Calculates the u-component of wind with forward motion added}
\usage{
calc_sfc_u(wind_sfc_sym, tcspd_u, swd, correction_factor)
}
\arguments{
\item{wind_sfc_sym}{A numeric vector with maximum 10-meter 1-minute
sustained wind with motion asymmetry removed (m / s).}

\item{tcspd_u}{A numeric vector with the tropical cyclone speed, u-component
(m / s).}

\item{swd}{A numeric vector with surface wind direction (in degrees).}

\item{correction_factor}{A numeric vector giving the correction factor,
calculated based on Phadke et al. 2003}
}
\description{
Adds the storm's forward speed component (i.e., motion asymmetery) back
into the estimated u-component of the surface wind speed at a grid point
location after u-component of rotational winds have been modeled for the
location.
}
\details{
This function uses equation 12 from Phadke et al. (2003).
}
\references{
Phadke AC, Martino CD, Cheung KF, and Houston SH. 2003. Modeling of
   tropical cyclone winds and waves for emergency management. Ocean
   Engineering 30(4):553-578.
}
