% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/my_wind_code.R
\name{calc_track}
\alias{calc_track}
\title{Holland2 model to calculate gradient windspeed distribution}
\usage{
calc_track(r, Rmax, known, r_vmax, n, A, X1)
}
\description{
Holland2 model to calculate gradient windspeed distribution
}
\note{
Wind speed scaled by 100 - must undo after calculation
}

