% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/my_wind_code.R
\name{calcangle}
\alias{calcangle}
\title{Apply the angle calculation across a vector}
\usage{
calcangle(dx, dy)
}
\arguments{
\item{dx}{Numeric vector with x-component of forward speed, in m / s}

\item{dy}{Numeric vector with y-component of forward speed, in m / s}
}
\value{
This function returns a vector with wind direction angles.
}
\description{
Apply the angle calculation across a vector
}

