% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helper.R
\name{calcangle_pair}
\alias{calcangle_pair}
\title{Calculate motion direction angle from ...}
\usage{
calcangle_pair(dx, dy)
}
\arguments{
\item{dpair}{Numeric vector of length 2 with the values for "dx" and "dy"}
}
\value{
This function returns the motion direction angle ...
}
\description{
Use trig angles, so E=0, N=90, W=180, S=270)
}

