% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/wind_conversions.R
\name{check_over_land}
\alias{check_over_land}
\title{Determine if storm is over land or water}
\usage{
check_over_land(tclat, tclon)
}
\arguments{
\item{tclat}{Numeric vector of the value of latitude, in degrees.}

\item{tclon}{Numeric vector of the absolute value of latitude, in degrees.}
}
\value{
A logical vector of whether the storm is over land (TRUE) or water
   (FALSE)
}
\description{
Determines if the storm is over land or water at its observed location. This
function finds the closest grid point in the \code{\link{landmask}} dataframe, then
checks if that grid point is over land or water.
}
