% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{find_x_section}
\alias{find_x_section}
\title{Find the location of a new value in a table of values}
\usage{
find_x_section(new_x, x)
}
\arguments{
\item{new_x}{A double floating point number giving a new value of x for which
you want to find the indices of the sections of a vector x that contain the
new value.}

\item{x}{A numeric vector for which you want to find the position that brackets
the new value, new_x.}
}
\value{
An integer giving the zero-based index of the position in the x vector
  that gives the lower bracket of the section of the x-vector containing the new
  x value.
}
\description{
Find the location of a new value in a table of values
}
