% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{gradient_to_surface}
\alias{gradient_to_surface}
\title{Calculate symmetrical surface wind from gradient wind}
\usage{
gradient_to_surface(wind_gl_aa, cdist, glandsea)
}
\arguments{
\item{wind_gl_aa}{The gradient-level wind at the location at which the wind
is being modeled (in m/s)}

\item{cdist}{The distance from the center of the storm to the location at
which the wind is being modeled (in kilometers)}

\item{glandsea}{A logical value that specifies whether the location at which
the wind is being modeled is over land (TRUE) or over water (FALSE)}
}
\value{
A numeric value giving the surface-level wind at that location (before
  incorporating the role of forward motion of the storm) (in m/s)
}
\description{
Calculates the surface wind speed based on an estimated gradient
wind speed at a point and the radius from the storm center to
the grid point.
}
\details{
The reduction factor is based on Figure 3 of Knaff et al., 2003.
Over water, it is estimated to be 0.9 up to a radius of 100 km,
0.75 for a radius of 700 km or more, and decrease linearly between
a radius of 100 km and 700 km. Points over land use a reduction
factor that is 20\% lower.
}
\references{
Knaff JA, DeMaria M, Molenar DA, Sampson CR, and Seybold MG. 2011. An
automated, objective, multiple-satellite-platform tropical cyclone surface
wind speed analysis. Journal of Applied Meteorology and Climatology
50(10):2149-2166
}
