% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{gradient_to_surface_new}
\alias{gradient_to_surface_new}
\title{Calculate symmetrical surface wind from gradient wind}
\usage{
gradient_to_surface_new(wind_gl_aa, cdist, glandsea)
}
\arguments{
\item{wind_gl_aa}{The gradient-level wind at the location at which the wind
is being modeled (in m/s)}

\item{cdist}{The distance from the center of the storm to the location at
which the wind is being modeled (in kilometers)}

\item{glandsea}{A logical value that specifies whether the location at which
the wind is being modeled is over land (TRUE) or over water (FALSE)}
}
\value{
A numeric value giving the surface-level wind at that location (before
  incorporating the role of forward motion of the storm) (in m/s)
}
\description{
Calculate symmetrical surface wind from gradient wind
}
\references{

}
