% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{interpolate_line}
\alias{interpolate_line}
\title{Interpolate new y-values using linear interpolation}
\usage{
interpolate_line(x, y, new_x)
}
\arguments{
\item{x}{A numeric vector with the x values of the original function.}

\item{y}{A numeric vector with the y values of the original function. This should
be the same length as x.}

\item{new_x}{A numeric value providing the new x values at which you want to
interpolate y values.}
}
\value{
A numeric vector with the interpolated values of y at each of the new
  x values provided by the new_x input.
}
\description{
Interpolate new y-values using linear interpolation
}
