% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{interpolate_spline}
\alias{interpolate_spline}
\title{Perform spline interpolation}
\usage{
interpolate_spline(x, y, new_x)
}
\arguments{
\item{x}{A numeric vector with independent values of the original function.}

\item{y}{A numeric vector with dependent values of the original function.}

\item{new_x}{A numeric vector with the values of x at which you want to interpolate v
values of y.}
}
\value{
A numeric vector with the interpolated values of y at the \code{new_x} values.
}
\description{
Perform spline interpolation
}
