% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{katrina_tracks}
\alias{katrina_tracks}
\title{Hurricane Katrina tracks data}
\format{
A dataframe with 48 rows and 4 variables:
\describe{
    \item{date}{A character string giving the date and time of the observation}
    \item{latitude}{A numeric vector giving the storm's latitude at that
                    observation time}
    \item{longitude}{A numeric vector giving the storm's longitude at that
                    observation time}
    \item{wind}{A numeric vector giving the estimated maximum sustained
                wind of that storm at that observation time, in knots}
}
}
\source{
\url{http://rammb.cira.colostate.edu/research/tropical_cyclones/tc_extended_best_track_dataset/}
}
\usage{
katrina_tracks
}
\description{
A dataframe containing hurricane best tracks for Hurricane Katrina in
2005. This dataframe is included for use as an example hurricane
tracks dataframe in the package documentation. This data originally
came from the Extended Best Track dataset:
\url{http://rammb.cira.colostate.edu/research/tropical_cyclones/tc_extended_best_track_dataset/}
}
\keyword{datasets}
