% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/latlon_helpers.R
\name{latlon_to_km}
\alias{latlon_to_km}
\title{Calculate distance between two locations}
\usage{
latlon_to_km(tclat_1, tclon_1, tclat_2, tclon_2)
}
\arguments{
\item{tclat_1}{Latitude of the first location, in degrees.}

\item{tclon_1}{Longitude of the first location, in degrees. This value should
be expressed as a positive value for Western hemisphere longitudes.}

\item{tclat_2}{Latitude of the second location, in degrees.}

\item{tclon_2}{Longitude of the second location, in degrees. This value should
be expressed as a positive value for Western hemisphere longitudes.}
}
\value{
Distance between the two locations, in kilometers.
}
\description{
This function takes latitude-longitude values for two locations and
calculates the distance (in meters) between the locations using the
haversine method.
}
\details{
This function uses the Haversine method with great circle distance
   to calculate this distance. It is applying the following equations to
   determine distance (in kilometers) between two latitude-longitude pairs:
   \deqn{hav(\theta) = hav(\phi_1 - \phi_2) + cos(\phi_1)*cos(\phi_2)*hav(L_1 - L_2)}{
   hav(\theta) = hav(\phi1 - \phi2) + cos(\phi1)*cos(\phi2)*hav(L1 - L2)}
   where:
   \itemize{
     \item{\eqn{\phi_1}{\phi1}: Latitude of first location, in radians}
     \item{\eqn{\phi_2}{\phi2}: Latitude of second location, in radians}
     \item{\eqn{L_1}{L1}: Longitude of first location, in radians}
     \item{\eqn{L_2}{L2}: Longitude of second location, in radians}
     \item{\eqn{hav(\alpha)}: The haversine function,
        \eqn{hav(\alpha) = sin^2 \left(\frac{\alpha}{2}\right)}{
        hav(\alpha) = sin^2 (\alpha / 2)}}
     \item{\eqn{R}: Radius of the earth, here assumed to be 6378.14 kilometers}
     \item{\eqn{D}}: Distance between the two locations, in kilometers
   }
}
\note{
This function assumes the radius of the earth is 6378.14 kilometers.
}

