% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/latlon_helpers.R
\name{latlon_to_meters}
\alias{latlon_to_meters}
\title{Calculate distance between two locations}
\usage{
latlon_to_meters(phi_1, L_1, phi_2, L_2)
}
\arguments{
\item{phi_1}{Latitude of the first location, in degrees}

\item{L_1}{Longitude of the first location, in degrees}

\item{phi_2}{Latitude of the second location, in degrees}

\item{L_2}{Longitude of the second location, in degrees}
}
\value{
Distance between the two locations, in kilometers.
}
\description{
This function takes latitude-longitude values for two locations and
calculates the distance (in meters) between the locations using the
haversine method.
}
\note{
This function assumes the radius of the earth is 6378.14 kilometers.
}

