% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/my_wind_code.R
\name{map_wind}
\alias{map_wind}
\title{Map wind exposure at the county level}
\usage{
map_wind(grid_winds, value = "maxwindspd", break_point = NULL)
}
\description{
Map wind exposure at the county level
}

