% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{saomai_tracks}
\alias{saomai_tracks}
\title{Typhoon Saomai tracks data}
\format{
A dataframe with 67 rows and 5 variables:
\describe{
    \item{iso_time}{A date-time value giving the time of the observation (in UTC)}
    \item{lat}{A numeric vector giving the storm's latitude at that
                    observation time}
    \item{lon}{A numeric vector giving the storm's longitude at that
                    observation time}
    \item{usa_wind}{A numeric vector giving the estimated maximum sustained
                wind of that storm at that observation time, in knots}
    \item{usa_rmw}{A numeric vector giving the estimated radius to maximum wind, in
                nautical miles.}
}
}
\source{
\url{https://www.ncei.noaa.gov/products/international-best-track-archive?name=sources}
}
\usage{
saomai_tracks
}
\description{
A dataframe containing tracking data for Typhoon Saomai in
2006. This dataframe is included for use as an example storm
tracks dataframe in the package documentation. This data originally
came from the International Best Track Archive for Climate Stewardship (IBTrACS):
\url{https://www.ncdc.noaa.gov/ibtracs/}
}
\keyword{datasets}
