% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/my_wind_code.R
\name{summarize_grid_wind}
\alias{summarize_grid_wind}
\title{Generate wind summaries for grid point}
\usage{
summarize_grid_wind(grid_wind, tint = 0.25, gust_duration_cut = 20,
  sust_duration_cut = 20)
}
\description{
Generate wind summaries for grid point
}

