% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/willoughby2006eqs.R
\name{will1}
\alias{will1}
\title{Model wind speed at each grid point for each storm track observation}
\usage{
will1(r, Rmax, R1, R2, Vmax, n, A, X1, X2 = 25)
}
\arguments{
\item{r}{Numeric vector of radius from the storm center to the point you are
measuring, in kilometers}

\item{Rmax}{Numeric vector of the radius at which the maximum wind occurs,
in kilometers}

\item{Vmax}{Numeric vector of the tangential wind component of the maximum
gradient wind speed, in meters per second}

\item{n}{Numeric vector of ...}

\item{A}{A numeric vector ...}

\item{X1}{A numeric vector ...}
}
\description{
Model wind speed at each grid point for each storm track observation
}

