% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/willoughby2006eqs.R
\name{will10a}
\alias{will10a}
\title{Calculate the fitted decay length from \eqn{Vmax} and \eqn{\phi}}
\usage{
will10a(Vmax, phi)
}
\arguments{
\item{Vmax}{Numeric vector of the tangential wind component of the maximum
gradient wind speed, in meters per second}

\item{phi}{Numeric vector of the absolute value of latitude, in degrees.}
}
\value{
A numeric vector of the fitted slower decay length for the dual
   exponential windspeed model (km, I think).
}
\description{
This function implements Willoughby et al. (2006), Equation 10a.
}

