% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/willoughby2006eqs.R
\name{will1a}
\alias{will1a}
\title{Willoughby et al. (2006), Equation 1(a)}
\usage{
will1a(vmax_gl, r, Rmax, n)
}
\arguments{
\item{vmax_gl}{Numeric vector of the tangential wind component of the maximum
gradient wind speed, in meters per second}

\item{r}{Numeric vector of radius from the storm center to the point you are
measuring, in kilometers}

\item{Rmax}{Numeric vector of the radius at which the maximum wind occurs,
in kilometers}

\item{n}{A numeric vector of one of the parameters of the Willoughby model.}
}
\description{
Willoughby et al. (2006), Equation 1(a)
}

