% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{will1new}
\alias{will1new}
\title{Calculate gradient wind speed using equation 1 from Willoughby}
\usage{
will1new(cdist, Rmax, R1, R2, vmax_gl, n, A, X1, X2 = 25)
}
\arguments{
\item{cdist}{A numeric value with the distance between the grid point and the
center of the storm (in km)}

\item{Rmax}{A numeric value with the radius at which the maximum wind occurs
(in km)}

\item{R1}{A numeric value with the lower boundary of the transition zone (in
km from the storm center)}

\item{R2}{A numeric value with the upper boundary of the transition zone (in
km from the storm center)}

\item{vmax_gl}{A numeric value with the maximum gradient level 1-minute
sustained wind, in (m/s)}

\item{A}{A numeric value as a parameter for the Willoughby model}

\item{X1}{A numeric value as a parameter for the Willoughby model}

\item{X2}{A numeric value as a parameter for the Willoughby model, set to 25
(Willoughby, Darling, and Rahn 2006)}

\item{r}{A numeric value with the radius from the storm center to the grid
point (in km)}
}
\value{
wind_gl_aa A numeric value ...
}
\description{
Calculate gradient wind speed using equation 1 from Willoughby
}
