% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/willoughby2006eqs.R
\name{will2}
\alias{will2}
\title{Willoughby et al. (2006), Equation 2}
\usage{
will2(r, R1, R2)
}
\arguments{
\item{r}{Numeric vector of radius from the storm center to the point you are
measuring, in kilometers}

\item{R1}{Numeric vector of radius at start of transition zone, in
kilometers}
}
\value{
A numeric vector of the weighting parameter for Willoughby's wind
   profile equations (\eqn{w}).
}
\description{
Calculates equation 2 from Willoughby et al. (2006).
}
