% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{will2_new}
\alias{will2_new}
\title{Calculate equation 2 from Willoughby}
\usage{
will2_new(r, R1, R2)
}
\arguments{
\item{r}{A numeric value giving the distance from the center of the storm to
the location where you would like to model the local wind, in kilometers}

\item{R1}{A numeric value with the lower boundary of the transition zone (in
km from the storm center)}

\item{R2}{A numeric value with the upper boundary of the transition zone (in
km from the storm center)}
}
\value{
A numeric value between 0 and 1 that gives the amount to weight Vi
  versus Vo when estimating wind within the transition zone.
}
\description{
Calculate equation 2 from Willoughby
}
