% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/willoughby2006eqs.R
\name{will4}
\alias{will4}
\title{Willoughby et al. (2006), Equation 4}
\usage{
will4(vmax_gl, A, r, Rmax, X1, X2 = 25)
}
\arguments{
\item{vmax_gl}{Numeric vector of the tangential wind component of the maximum
gradient wind speed, in meters per second}

\item{A}{A numeric vector of one of the parameters of the Willoughby model.}

\item{r}{Numeric vector of radius from the storm center to the point you are
measuring, in kilometers}

\item{Rmax}{A numeric vector giving the radius to maximum winds (in kilometers)
for the tropical storm.}

\item{X1}{A numeric vector of one of the parameters of the Willoughby model.}
}
\description{
Willoughby et al. (2006), Equation 4
}
