% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/willoughby2006eqs.R
\name{will7a}
\alias{will7a}
\title{Calculate Rmax from Vmax and phi}
\usage{
will7a(Vmax, phi)
}
\arguments{
\item{Vmax}{Numeric vector of the tangential wind component of the maximum
gradient wind speed, in meters per second}

\item{phi}{Numeric vector of the absolute value of latitude, in degrees.}
}
\value{
A numeric vector with the radius at which the maximum wind occurs,
   in kilometers.
}
\description{
This function calculates the radius at which the maximum wind occurs
(\eqn{Rmax}), based on the tangential wind component of the maximum wind
speed (\eqn{Vmax}) and latitude (\eqn{\phi}). This function implements
Willoughby et al. (2006), Equation 7a.
}

