% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R, R/boundary_layer_helpers.R
\name{add_forward_speed}
\alias{add_forward_speed}
\title{Add in forward speed of the storm}
\usage{
add_forward_speed(wind_sfc_sym, tcspd_u, tcspd_v, swd, cdist, Rmax)

add_forward_speed(wind_sfc_sym, tcspd_u, tcspd_v, swd, cdist, Rmax)
}
\arguments{
\item{wind_sfc_sym}{A numeric vector with maximum 10-meter 1-minute
sustained wind with motion asymmetry removed (m / s).}

\item{tcspd_u}{A numeric vector with the tropical cyclone speed, u-component
(m / s).}

\item{tcspd_v}{A numeric vector with the tropical cyclone speed, v-component
(m / s).}

\item{swd}{A numeric vector with surface wind direction (in degrees).}

\item{cdist}{A numeric value with the radius from the storm's center to the
grid point in kilometers}

\item{Rmax}{A numeric value with the radius at which maximum winds occur in kilometers}
}
\value{
windspeed A numeric value with the asymmetric surface windspeed at the
location, in meters/second

A numeric vector giving asymmetric surface windspeed (m / s) at the
   location being modeled.
}
\description{
Adds the storm's forward speed component (i.e., motion asymmetry) back
into the estimated surface wind speed at a grid point location after
rotational winds have been modeled for the location.

Adds the storm's forward speed component (i.e., motion asymmetry) back
into the estimated surface wind speed at a grid point location after
rotational winds have been modeled for the location.
}
\details{
This function uses equation 12 from Phadke et al. (2003).

This function uses equation 12 from Phadke et al. (2003).
}
\references{
Phadke AC, Martino CD, Cheung KF, and Houston SH. 2003. Modeling of
   tropical cyclone winds and waves for emergency management. Ocean
   Engineering 30(4):553-578.

Phadke AC, Martino CD, Cheung KF, and Houston SH. 2003. Modeling of
   tropical cyclone winds and waves for emergency management. Ocean
   Engineering 30(4):553-578.
}
