% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/boundary_layer_helpers.R
\name{add_forward_speed}
\alias{add_forward_speed}
\title{Adds forward speed component}
\usage{
add_forward_speed(windspd, forward_speed_u, forward_speed_v, swd, r, Rmax)
}
\description{
Adds the storm's forward speed component back into the estimated
surface wind speed.
}

