% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/boundary_layer_helpers.R
\name{add_forward_speed}
\alias{add_forward_speed}
\title{Calculate total modeled surface wind speed}
\usage{
add_forward_speed(wind_sfc_u, wind_sfc_v)
}
\arguments{
\item{wind_sfc_u}{A numeric vector with the tropical cyclone speed, u-component
(m / s).}

\item{wind_sfc_u}{A numeric vector with the tropical cyclone speed, v-component
(m / s).}
}
\value{
A numeric vector giving asymmeric surface windspeed (m / s) at the
   location being modeled.
}
\description{
Adds together the u- and v-components of the surface winds, after the
forward motion of the storm has been added back in, to calculate the
total windspeed at a grid point.
}
