% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/my_wind_code.R
\name{add_wind_radii}
\alias{add_wind_radii}
\title{Calculate Willoughby parameters for the storm}
\usage{
add_wind_radii(full_track = create_full_track())
}
\arguments{
\item{full_track}{An interpolated version of the hurricane track data, as
created by \code{create_full_track}. This dataframe must have the
following columns:
\itemize{
  \item{date: Date-time, in POSIXct format and UTC time zone}
  \item{tclat: Latitude, in decimal degrees}
  \item{tclon: Longitude, in decimal degrees and expressed as positive
        values (this model assumes all longitudes will be in the
        Western hemisphere)}
  \item{vmax: Sustained maximum wind speed, in meters per
       second}
}}
}
\value{
The input dataset, but with columns added for the Willoughby
   parameters for every observations. Columns added include:
   \itemize{
     \item{tcspd: The forward (translational) speed of the storm,
       in meters per second}
     \item{tcdir: The bearing of the storm, in degrees, with 0 degrees
       indicating the storm is moving due east, 90 degrees indicating the
       storm is moving due north, etc.}
     \item{Vmax: The gradient wind speed, based on the best track's
       sustained surface wind speed}
     \item{Rmax: The radius to maximum winds, in kilometers}
     \item{X1, n, A, eq3_right, xi, R1, R2: Different parameters needed for
       the Willoughby model}
   }
}
\description{
There are a number of parameters needed for the Willoughby model for each
observation time for a storm. This function intakes the storm track for
the storm, interpolated first if you want, and calculates these parameters
for each observation in the storm tracks, based on the latitude and longitude
of the storm at that observation, its forward speed and direction, and
its maximum windspeed.These parameters are later used to model wind speed
at each grid point for every observation.
}
\examples{
\dontrun{
data("floyd_tracks")
full_track <- create_full_track(hurr_track = floyd_tracks)
with_wind_radii <- add_wind_radii(full_track = full_track)
head(with_wind_radii)
}

}

