% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/latlon_helpers.R
\name{calc_bearing}
\alias{calc_bearing}
\title{Calculate the direction from one location to another}
\usage{
calc_bearing(tclat_1, tclon_1, tclat_2, tclon_2)
}
\arguments{
\item{tclat_1}{Latitude of the first location, in degrees.}

\item{tclon_1}{Longitude of the first location, in degrees. This value should
be expressed as a positive value for Western hemisphere longitudes.}

\item{tclat_2}{Latitude of the second location, in degrees.}

\item{tclon_2}{Longitude of the second location, in degrees. This value should
be expressed as a positive value for Western hemisphere longitudes.}
}
\value{
The direction of the second location as seen from the first location,
   in degrees. A direction of 0 degrees indicates the second location is
   due east of the first, and 90 degrees if the second location is due north
   of the first.
}
\description{
This function calculates the bearing of a second location, as seen from
the first location, based on latitude and longitude coordinates for both
locations.
}
\details{
This function uses the following equations to calculate the bearing
   from one latitude-longitude pair to another:

   \deqn{S = cos(\phi_2) * sin(L_1 - L_2)}{
   S = cos(\phi2) * sin(L1 - L1)}

   \deqn{C = cos(\phi_1) * sin(\phi_2) - sin(\phi_1) * cos(\phi_2) * cos(L_1 - L_2)}{
   C = cos(\phi1) * sin(\phi2) - sin(\phi1) * cos(\phi2) * cos(L1 - L2)}

   \deqn{\theta = atan2(S, C) * \frac{180}{\pi} + 90}

   where:
   \itemize{
     \item{\eqn{\phi_1}{\phi1}: Latitude of first location, in radians}
     \item{\eqn{L_1}{L1}: Longitude of first location, in radians}
     \item{\eqn{\phi_2}{\phi2}: Latitude of second location, in radians}
     \item{\eqn{L_2}{L2}: Longitude of second location, in radians}
     \item{\eqn{S, C}: Intermediary results}
     \item{\eqn{\theta}: Direction of the storm movement, in degrees}
   }

   In cases where this equation results in values below 0 degrees or above
   360 degrees, the function applies modular arithmetic to bring the value
   back within the 0-360 degree range.
}

