% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/latlon_helpers.R
\name{calc_forward_speed}
\alias{calc_forward_speed}
\title{Calculate storm's forward speed}
\usage{
calc_forward_speed(phi_1, L_1, time_1, phi_2, L_2, time_2)
}
\arguments{
\item{phi_1}{Latitude of the first location, in degrees}

\item{L_1}{Longitude of the first location, in degrees}

\item{time_1}{Time of the first observation.}

\item{phi_2}{Latitude of the second location, in degrees}

\item{L_2}{Longitude of the second location, in degrees}

\item{time_2}{Time of the second observation.}
}
\value{
A numeric vector with the average forward speed of the storm between
   the two observations, in meters per second.
}
\description{
This storm takes two latitude-longitude measurements, and the time of
each measurement, and calculates the average speed of the storm between the
two observations.
}

