% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/wind_conversions.R
\name{calc_gradient_speed}
\alias{calc_gradient_speed}
\title{Convert sustained wind to gradient wind.}
\usage{
calc_gradient_speed(sustained_vmax)
}
\arguments{
\item{sustained_vmax}{A numeric vector of 1-minute sustained wind speed at
10 meters, in meters / second.}
}
\value{
A numeric vector with gradient level wind speed, in meters / second.
}
\description{
This function converts 1-minute sustained wind speed at 10 meters to gradient
level wind speed.
}

