% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/wind_conversions.R
\name{calc_gradient_speed}
\alias{calc_gradient_speed}
\title{Convert sustained wind to gradient wind.}
\usage{
calc_gradient_speed(sustained_vmax, over_land)
}
\arguments{
\item{sustained_vmax}{A numeric vector of 1-minute sustained wind speed at
10 meters, in meters / second.}

\item{over_land}{TRUE / FALSE or whether the storm is over land (the
alternative is that the storm is over water).}
}
\value{
A numeric vector with gradient-level wind speed, in meters / second.
}
\description{
This function converts 1-minute sustained wind speed at 10 meters to gradient
level wind speed.
}
\details{
This function uses the following conversion:
 \deqn{V_{max} = \frac{V_{max,sustained}}{0.9}}{
 Vmax = Vmax_sustained / reduction_factor}
 where:
 \itemize{
   \item{\eqn{V_{max}}{Vmax}: Mean wind speed at gradient level (m / s) }
   \item{\eqn{V_{max,sustained}}{Vmax_sustained}: Surface wind speed (10 meters above the water or ground) (m / s)}
 }
 We make this adjustment based on Figure 3 in Knaff et al., 2011.
   If over water and within 100 kilometers of the storm's center,
   the ratio of gradient wind speed to surface wind speed is
   assumed to be 0.90. If over land, this reduction factor is reduced
   by 20\% (\eqn{0.9 * 0.8 = 0.72}).

   For this calculation, we assume
   that the radius of
   maximum wind for all storms is 100 kilometers or smaller (in the code,
   we do not calculate \eqn{R_{max}}{R_max} until after we estimate
   gradient wind speed from surface wind speed, so we don't have that
   storm-specific estimate to use here).
}

