% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/boundary_layer_helpers.R
\name{calc_gwd}
\alias{calc_gwd}
\title{Calculate the gradient wind direction}
\usage{
calc_gwd(tclat, tclon, glat, glon)
}
\value{
Numeric vector with the gradient wind direction (in degrees),
prior to adding an inflow angle.
}
\description{
This function determines the bearing from the storm's center to a
grid point location and then calculates the direction of gradient
winds at that grid point. Gradient winds will be perpendicular to
the direction from the storm's center to the grid point. In the
Northern hemisphere, cyclonic winds are counterclockwise, so you
add 90 degrees to the bearing from the storm center to the grid point
when using polar coordinates. In the Southern hemisphere, cyclonic
winds are clockwise, so you subtract 90 degrees.
}
\examples{
calc_gwd(tclat = 41.4, tclon = -72.8,
         glat = 32.50039, glon = -86.49416)

calc_gwd(tclat = -41.4, tclon = -72.8,
         glat = -32.50039, glon = -86.49416)

}
