% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{degrees_to_radians_Cpp}
\alias{degrees_to_radians_Cpp}
\title{Convert from degrees to radians}
\usage{
degrees_to_radians_Cpp(degrees)
}
\arguments{
\item{degrees}{A numeric vector with measurements in degrees.}
}
\value{
A numeric vector with measurement in radians.
}
\description{
Convert an angle from degrees to radians (C++ version)
}
