% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/wind_conversions.R
\name{gradient_to_surface}
\alias{gradient_to_surface}
\title{Calculate surface wind speed from gradient}
\usage{
gradient_to_surface(track, r)
}
\arguments{
\item{track}{Estimated gradient-level wind speed (m / s) at a grid
point.}

\item{r}{Radius from storm center to the grid point, in kilometers.}
}
\value{
Estimate of surface wind speed at grid point, in meters / second.
}
\description{
Calculates the surface wind speed based on an estimated gradient
wind speed at a point and the radius from the storm center to
the grid point.
}
\details{
The reduction factor is based on Figure 3 of Knaff et al., 2003.
It is estimated to be, for over water, 0.9 up to a radius of 100 km,
0.75 for a radius of 700 km or more, and decrease linearly between
a radius of 100 km and 700 km. Points over land should use a reduction
factor that is 20% lower. Because all of the counties are over
land, we make this adjustment for all grid points.
}

