% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/map_winds.R
\name{map_wind}
\alias{map_wind}
\title{Map wind exposure at the county level}
\usage{
map_wind(grid_winds, value = "vmax_sust", break_point = NULL,
  wind_metric = "mps")
}
\arguments{
\item{grid_winds}{A dataframe that is the output of
\code{\link{get_grid_winds}}.}

\item{value}{A character string giving the value to plot. Possible options
are "max_gust" (maximum gust wind speeds) and "max_sust" (maximum
sustained wind speeds).}

\item{break_point}{An numeric value giving the value of the "value"
parameter to break at for a binary map showing exposure versus no exposure.}

\item{wind_metric}{A character vector with the wind metric to use for the map.
Possible values include \code{"knots"}, \code{"mph"}, \code{"ftps"} (ft / s),
\code{"kmph"} (km / hr), and \code{"mps"} (m / s, the default).}
}
\value{
This function returns a map of the "ggplot" class, plotting
   exposure to hurricane winds by county for the eastern half of the United
   States.
}
\description{
Map wind exposure at the county level
}
\examples{
\donttest{
data("katrina_tracks")
data("county_point")
grid_winds_katrina <- get_grid_winds(hurr_track = katrina_tracks,
                                     grid_df = county_points)
map_wind(grid_winds_katrina)
map_wind(grid_winds_katrina, break_point = 20)
}

}

