% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/map_winds.R
\name{map_wind}
\alias{map_wind}
\title{Map wind exposure at the county level}
\usage{
map_wind(grid_winds, value = "max_gust", break_point = NULL)
}
\arguments{
\item{grid_winds}{A dataframe that is the output of
\code{\link{get_grid_winds}}.}

\item{value}{A character string giving the value to plot. Possible options
are "maxwindspd" and "duration".}

\item{break_point}{An numeric value giving the value of the "value"
parameter to break at for a binary map showing exposure versus no exposure.}
}
\value{
This function returns a map of the "ggplot" class, plotting
   exposure to hurricane winds by county for the eastern half of the United
   States.
}
\description{
Map wind exposure at the county level
}
\examples{
\donttest{
load("writing/grid_winds_katrina.Rdata")
map_wind(grid_winds_katrina)
map_wind(grid_winds_katrina, break_point = 20)
}

}

