% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/wind_conversions.R
\name{remove_forward_speed}
\alias{remove_forward_speed}
\title{Remove forward speed estimate from maximum wind speed}
\usage{
remove_forward_speed(vmax, tcspd)
}
\value{
A numerical vector with the maximum storm wind speed, with forward
storm motion speed removed, in m / s.
}
\description{
This function takes the forward speed of the storm and subtracts it from
the maximum storm wind speed, \eqn{V_{max}}.
}
\details{
This function is based on equation 12 (and accompanying
   text) in Phadke et al. 2003. Based on this paper, the correction
   factor for forward speed is at its highest at the radius of
   maximum winds, where it equals 0.5 times the forward wind speed.
}
\references{
Padke AC, Martino CD, Cheung KF, and Houston SH. 2003. Modeling of
   tropical cyclone winds and waves for emergency management. Ocean
   Engineering 30:553-578.
}

