% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/willoughby2006eqs.R
\name{will10b}
\alias{will10b}
\title{Calculate the power law exponential from \eqn{vmax_gl} and \eqn{\phi}}
\usage{
will10b(vmax_gl, tclat)
}
\arguments{
\item{vmax_gl}{Numeric vector of the tangential wind component of the maximum
gradient wind speed, in meters per second}

\item{tclat}{Numeric vector of the absolute value of latitude, in degrees.}
}
\value{
A numeric vector of the exponential for the power law inside the eye
}
\description{
This function implements Willoughby et al. (2006), Equation 10b.
}
\details{
This function is calculating the equation:
   \deqn{n = 0.4067 + 0.0144 V_{max} - 0.0038 \phi}{
   n = 0.4067 + 0.0144 vmax_gl - 0.0038 \phi}
   where:
   \itemize{
     \item{\eqn{n}: Parameter for the Willoughby wind model}
     \item{\eqn{V_{max}}{vmax_gl}: Tangential component of the maximum
           gradient-level sustained wind speed (in m / s)}
     \item{\eqn{\phi}: Latitude, in decimal degrees}
   }
}

