% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/willoughby2006eqs.R
\name{will7a}
\alias{will7a}
\title{Calculate Rmax from Vmax and phi}
\usage{
will7a(Vmax, phi)
}
\arguments{
\item{Vmax}{Numeric vector of the tangential wind component of the maximum
gradient wind speed, in meters per second}

\item{phi}{Numeric vector of the absolute value of latitude, in degrees.}
}
\value{
A numeric vector with the radius at which the maximum wind occurs,
   in kilometers.
}
\description{
This function calculates the radius at which the maximum wind occurs
(\eqn{Rmax}), based on the tangential wind component of the maximum wind
speed (\eqn{Vmax}) and latitude (\eqn{\phi}). This function implements
Willoughby et al. (2006), Equation 7a.
}
\details{
This function fits the following equation:
\deqn{R_{max} = 46.4 e^{- 0.0155 V_{max} + 0.0169\phi}}{
Rmax = 46.4 e^(- 0.0155 Vmax) + 0.0169\phi}
where:
\itemize{
  \item{\eqn{R_{max}}{Rmax}: Radius from the storm center to the point at which the maximum wind occurs (km)}
  \item{\eqn{V_{max}}{Vmax}: Tangential wind component of the gradient-level maximum wind speed (m / s)}
  \item{\eqn{\phi}: Latitude (degrees)}
}
}

